Minor fix, move some functions into pid.ts
