+++ l2/eigen_service/src/login/config.ts
@@ -4,7 +4,7 @@ export const GOOGLE_CLIENT_ID =
 export const GOOGLE_CLIENT_SECRET =
   process.env.GOOGLE_CLIENT_SECRET || "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";
 
-export const UI_ROOT_URI = "https://secret.ieigen.com";
+export const UI_ROOT_URI = "https://secret.ieigen.com/home";
 export const SERVER_ROOT_URI = "https://rpc.ieigen.com/api";
 export const JWT_SECRET = "import-secret-from-env"; //default hmac sha256
